Create render.php
This is a drop-in template replacement for Bare which adds language translations and custom input forms for other plugins. The language file comes from the  LANGUAGE configuration in Bare (also as 'language' in config.json, if given). E.G. if the default language is 'en-US', the render plugin will try to load 'en-US.json' from the /cache directory.

Over time, most custom rendering functions will be moved to the render plugin with the core Bare blog maintaining just the basics.